Set the outdated flag
It is a first attempt to fix this part but I just discovered how the
code works (partially) and I now think the whole version comparison and
displayed info should be reworked. Upstream's issues and PR show that
people don't get what "outdated" means and it's not that useful as it
is.